agent_system/pyproject.toml
[project]
name = "agent-system"
version = "0.1.0"
description = "通用Agent系统"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0.0",
    "sqlalchemy>=2.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "toml>=0.10.2",
    "asyncio",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
]